refactor(warnings): split glyph text detectors
The warnings orchestrator had grown into a mix of detector routing, glyph-map heuristics, and visual evidence checks. Moving glyph and text-noise detectors into their own module keeps warning domains easier to review without changing the emitted warning behavior.
